Move Node to inner class of StaticHierarchy
